refact(ProductItem): make ProductItemV2 more reusable
