+++ apps/store/src/components/ProductItemV2/ProductDetails.css.ts
