+++ apps/store/src/components/ProductItemV2/ProductDetails.tsx
