+++ apps/store/src/components/ProductItemV2/ProductItem.types.ts
